Move BaseStudyQueryView, DefaultRequestNotification, ParticipantDataset, SpecimenEventQueryView, SpecimenQueryView, SpecimenRequestDisplayColumn, SpecimenSearchBean, SpecimenSearchWebPart, and SpecimenSearchWebPartFactory
