+++ specimen/src/org/labkey/specimen/view/SpecimenSearchWebPartFactory.java
@@ -1,14 +1,14 @@
-package org.labkey.study;
+package org.labkey.specimen.view;
 
 import org.jetbrains.annotations.NotNull;
 import org.labkey.api.security.permissions.ReadPermission;
+import org.labkey.api.specimen.SpecimenSearchWebPart;
 import org.labkey.api.study.StudyService;
 import org.labkey.api.view.DefaultWebPartFactory;
 import org.labkey.api.view.HtmlView;
 import org.labkey.api.view.Portal;
 import org.labkey.api.view.ViewContext;
 import org.labkey.api.view.WebPartView;
-import org.labkey.study.specimen.SpecimenSearchWebPart;
 
 public class SpecimenSearchWebPartFactory extends DefaultWebPartFactory
 {
