+++ study/src/org/labkey/study/specimen/report/specimentype/TypeParticipantReportFactory.java
@@ -18,6 +18,7 @@ package org.labkey.study.specimen.report.specimentype;
 import org.labkey.api.data.SimpleFilter;
 import org.labkey.api.query.FieldKey;
 import org.labkey.api.specimen.SpecimenManager;
+import org.labkey.api.study.Cohort;
 import org.labkey.api.study.Study;
 import org.labkey.api.study.StudyService;
 import org.labkey.api.study.Visit;
@@ -25,10 +26,8 @@ import org.labkey.api.util.DemoMode;
 import org.labkey.api.util.HtmlString;
 import org.labkey.api.util.Pair;
 import org.labkey.study.controllers.specimen.SpecimenController;
-import org.labkey.study.model.CohortImpl;
 import org.labkey.study.model.Participant;
 import org.labkey.study.model.StudyManager;
-import org.labkey.study.model.VisitImpl;
 import org.labkey.study.specimen.report.SpecimenTypeVisitReport;
 import org.labkey.study.specimen.report.SpecimenVisitReport;
 
@@ -87,7 +86,7 @@ public class TypeParticipantReportFactory extends TypeReportFactory
         String[] participantIds;
         if (!isAllSubjectsOption(_participantId) && _participantId != null && _participantId.trim().length() > 0)
         {
-            Study study = StudyManager.getInstance().getStudy(getContainer());
+            Study study = StudyService.get().getStudy(getContainer());
             Participant participant = StudyManager.getInstance().getParticipant(study, _participantId);
             if (participant == null)
                 return Collections.emptyList();
@@ -95,7 +94,7 @@ public class TypeParticipantReportFactory extends TypeReportFactory
         }
         else
         {
-            Study study = StudyManager.getInstance().getStudy(getContainer());
+            Study study = StudyService.get().getStudy(getContainer());
             if (getParticipantGroupFilter() >= 0)
                 participantIds = StudyManager.getInstance().getParticipantIdsForGroup(study, getUser(), getParticipantGroupFilter());
             else
@@ -106,8 +105,8 @@ public class TypeParticipantReportFactory extends TypeReportFactory
         List<SpecimenVisitReport> reports = new ArrayList<>();
         Map<Integer, List<? extends Visit>> visitListCache = new HashMap<>(); // cohort rowId -> visits
         boolean showCohorts = StudyService.get().showCohorts(getContainer(), getUser());
-        List<VisitImpl> allVisits = null;
-        Study study = StudyManager.getInstance().getStudy(getContainer());
+        List<? extends Visit> allVisits = null;
+        Study study = StudyService.get().getStudy(getContainer());
         for (String participantId : participantIds)
         {
             SimpleFilter filter = new SimpleFilter(FieldKey.fromParts(StudyService.get().getSubjectColumnName(getContainer())), participantId);
@@ -115,7 +114,7 @@ public class TypeParticipantReportFactory extends TypeReportFactory
             List<? extends Visit> visits = null;
             if (showCohorts)
             {
-                CohortImpl cohort = StudyManager.getInstance().getCurrentCohortForParticipant(getContainer(), getUser(), participantId);
+                Cohort cohort = StudyManager.getInstance().getCurrentCohortForParticipant(getContainer(), getUser(), participantId);
                 if (cohort != null)
                 {
                     visits = visitListCache.get(cohort.getRowId());
